Restructure Scraper and add more test data
